Implement getProject
