+++ src/db/BadgeHubDataPostgresAdapter.ts
@@ -11,32 +11,22 @@ import { AppCategoryName } from "@domain/readModels/app/Category";
 import { Pool } from "pg";
 import { getPool } from "@db/connectionPool";
 import { DBProject as DBProject } from "@db/models/app/DBProject";
-import { DBVersion as DBVersion } from "@db/models/app/DBVersion";
 import sql, { join, raw } from "sql-template-tag";
-import { DBAppMetadataJSON as DBMetadataFileContents } from "@db/models/app/DBAppMetadataJSON";
-import { DBInsertUser, DBUser } from "@db/models/app/DBUser";
-import moment from "moment";
+import { DBInsertUser } from "@db/models/app/DBUser";
 import { getEntriesWithDefinedValues } from "@util/objectEntries";
 import { DBBadge } from "@db/models/DBBadge";
-import { DBDatedData } from "@db/models/app/DBDatedData";
-import { DatedData } from "@domain/readModels/app/DatedData";
+import {
+  getBaseSelectProjectQuery,
+  ProjectQueryResponse,
+  projectQueryResponseToReadModel,
+} from "@db/sqlHelpers/projectQuery";
+import { dateStringsToDates } from "@db/sqlHelpers/datedData";
 
 function getInsertKeysAndValuesSql(user: Object) {
   const definedEntries = getEntriesWithDefinedValues(user);
   const keys = join(definedEntries.map(([key]) => raw(key))); // raw is ok here because these keys are checked against our typescript definitions by tsoa
   const values = join(definedEntries.map(([, value]) => value));
   return { keys, values };
-}
-
-function dateStringsToDates(dbDatedData: DBDatedData): DatedData {
-  const datedData: DatedData = {
-    created_at: moment(dbDatedData.created_at).toDate(),
-    updated_at: moment(dbDatedData.updated_at).toDate(),
-  };
-  if (dbDatedData.deleted_at) {
-    datedData.deleted_at = moment(dbDatedData.deleted_at).toDate();
-  }
-  return datedData;
 }
 
 export class BadgeHubDataPostgresAdapter implements BadgeHubDataPort {
@@ -115,8 +105,11 @@ export class BadgeHubDataPostgresAdapter implements BadgeHubDataPort {
     );
   }
 
-  getProject(projectSlug: string): Promise<Project> {
-    throw new Error("Method not implemented.");
+  async getProject(projectSlug: string): Promise<Project> {
+    const project: ProjectQueryResponse = await this.pool
+      .query(sql`${getBaseSelectProjectQuery()} and p.slug = ${projectSlug}`)
+      .then((res) => res.rows[0]);
+    return projectQueryResponseToReadModel(project);
   }
 
   getVersion(projectSlug: string): Promise<Version> {
@@ -166,31 +159,7 @@ export class BadgeHubDataPostgresAdapter implements BadgeHubDataPort {
     badgeSlug?: string;
     appCategory?: AppCategoryName;
   }): Promise<Project[]> {
-    let query = sql`select p.slug,
-                           p.git,
-                           p.allow_team_fixes,
-                           p.user_id,
-                           p.created_at,
-                           p.updated_at,
-                           p.deleted_at,
-                           v.semantic_version,
-                           v.git_commit_id,
-                           v.published_at,
-                           v.revision,
-                           v.size_of_zip,
-                           m.description,
-                           m.interpreter,
-                           m.license_file,
-                           m.name,
-                           u.name as author_name
-                    from projects p
-                             left join users u on p.user_id = u.id
-                             left join versions v on p.version_id = v.id
-                             left join app_metadata_jsons m on v.app_metadata_json_id = m.id
-                    where p.deleted_at is null
-                      and u.deleted_at is null
-                      and v.deleted_at is null
-    `;
+    let query = getBaseSelectProjectQuery();
     if (filter?.pageLength) {
       query = sql`${query}
       limit
@@ -198,40 +167,9 @@ export class BadgeHubDataPostgresAdapter implements BadgeHubDataPort {
       offset
       ${filter?.pageStart ?? 0}`;
     }
-    const projects: (DBProject &
-      DBVersion &
-      DBMetadataFileContents & {
-        author_name: DBUser["name"];
-      })[] = await this.pool.query(query).then((res) => res.rows);
-    return projects.map((dbProject): Project => {
-      return {
-        version: undefined, // TODO
-        allow_team_fixes: false,
-        user_id: dbProject.user_id,
-        author: dbProject.author_name, // todo maybe change to email, full id or object with multiple fields
-        badges: [], // TODO
-        category: dbProject.category || "Uncategorised",
-        collaborators: [], // TODO
-        description: dbProject.description,
-        download_counter: undefined, // TODO
-        git: dbProject.git,
-        git_commit_id: dbProject.git_commit_id,
-        interpreter: dbProject.interpreter,
-        license: dbProject.license_file, // TODO check what we should do with the license, possibly we could say that this is either a path or 'MIT'|..., but then still we should read out the licens somewhere if it is a file.
-        name: dbProject.name,
-        published_at: moment(dbProject.published_at).toDate(),
-        revision: dbProject.revision,
-        size_of_content: undefined, // TODO
-        size_of_zip: dbProject.size_of_zip,
-        slug: dbProject.slug,
-        states: undefined,
-        status: undefined, // TODO
-        versions: undefined, // TODO
-        dependencies: undefined, // TODO
-        votes: undefined, // TODO
-        warnings: undefined, // TODO
-        ...dateStringsToDates(dbProject),
-      };
-    });
+    const projects: ProjectQueryResponse[] = await this.pool
+      .query(query)
+      .then((res) => res.rows);
+    return projects.map(projectQueryResponseToReadModel);
   }
 }
